fix : main page Scr 동작 구현완료
