+++ package.json
@@ -12,6 +12,7 @@
     "@emotion/styled": "^11.1.5",
     "@fullpage/react-fullpage": "^0.1.19",
     "@types/axios": "^0.14.0",
+    "@types/facepaint": "^1.2.1",
     "@types/recoil": "^0.0.1",
     "axios": "^0.21.1",
     "emotion-reset": "^3.0.0",
