+++ src/components/Main/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { MainSection, StartBtn, DownBtn } from './styles'
+import { MainSection, StartBtn } from './styles'
 import Image from 'next/image'
 import Link from 'next/link'
 
@@ -9,9 +9,10 @@ type Props = {
   desc: string
   background: string
   className: string
+  children: unknown
 }
 
-const Main = ({ title, desc, background }: Props) => {
+const Main = ({ title, desc, background, children }: Props) => {
   return (
     <MainSection className="section">
       <Image src={background} alt="믹챠 메인페이지 배경사진" layout="fill" objectFit="cover" />
@@ -21,10 +22,7 @@ const Main = ({ title, desc, background }: Props) => {
       <Link href="/start">
         <StartBtn type="button">2주 무료 이용 시작</StartBtn>
       </Link>
-
-      <DownBtn type="button">
-        <img src="/images/main_down.svg" alt="다음 섹션으로 이동하기" />
-      </DownBtn>
+      {children}
     </MainSection>
   )
 }
